refact(e2e): Move Dynamic Host Catalog env to its own file
This keeps each scenario consistent (an env_test.go file in each directory)
